Add AWS Text
Request was made to text to AWS: "The RESF utilizes AWS to run much of
the underlying infrastructure critical to the development and deployment
of the services supporting Rocky Linux. From Koji builders for x86_64
and aarch64 running in EC2 to SRPM blob storage in S3, AWS is
foundational to much of what we do."

This was attempted. The AWS icon was once cetered and smaller, but now
is larger and left justified. Please make sure this is what you want,
and if not, feel free to reject this PR. Admittedly, I'm a noob at this
stuff.